Compatibility with 32 bit boards

+++ src/SparkFun_ADXL345.cpp
@@ -74,9 +74,9 @@ void ADXL345::readAccel(int *x, int *y, int *z) {
 	readFrom(ADXL345_DATAX0, ADXL345_TO_READ, _buff);	// Read Accel Data from ADXL345
 	
 	// Each Axis @ All g Ranges: 10 Bit Resolution (2 Bytes)
-	*x = (((int)_buff[1]) << 8) | _buff[0];   
-	*y = (((int)_buff[3]) << 8) | _buff[2];
-	*z = (((int)_buff[5]) << 8) | _buff[4];
+	*x = (int16_t)((((int)_buff[1]) << 8) | _buff[0]);
+	*y = (int16_t)((((int)_buff[3]) << 8) | _buff[2]);
+	*z = (int16_t)((((int)_buff[5]) << 8) | _buff[4]);
 }
 
 void ADXL345::get_Gxyz(double *xyz){
